Move fee to frontend folder
